+++ CHANGELOG.md
@@ -188,7 +188,7 @@
 
 #### Behavior Change
 
-- When performing `DataFrame.set_index`, or setting `DataFrame.index` or `Series.index` where the length of the `Series`/`DataFrame` object does not match with the new index's length, a `ValueError` is no longer raised. When the `Series`/`DataFrame` object is longer than the new index, the `Series`/`DataFrame`'s new index is filled with `NaN` values for the "extra" elements. When the `Series`/`DataFrame` object is shorter than the new index, the extra values in the new index are ignored—`Series` and `DataFrame` stay the same length `n`, and use only the first `n` values of the new index.
+- When calling `DataFrame.set_index`, or setting `DataFrame.index` or `Series.index`, with a new index that does match the current length of the `Series`/`DataFrame` object does not match with the new index's length, a `ValueError` is no longer raised. When the `Series`/`DataFrame` object is longer than the new index, the `Series`/`DataFrame`'s new index is filled with `NaN` values for the "extra" elements. When the `Series`/`DataFrame` object is shorter than the new index, the extra values in the new index are ignored—`Series` and `DataFrame` stay the same length `n`, and use only the first `n` values of the new index.
 
 ## 1.20.0 (2024-07-17)